Move VersionedDao into the new storage package
