+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/dynamo/VersionedDao.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.dynamo
+package uk.ac.wellcome.storage.dynamo
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.google.inject.Inject
