+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/vhs/VersionedHybridStore.scala
@@ -3,9 +3,10 @@ package uk.ac.wellcome.storage.vhs
 import com.google.inject.Inject
 import com.gu.scanamo.DynamoFormat
 import io.circe.{Decoder, Encoder}
-import uk.ac.wellcome.dynamo.{UpdateExpressionGenerator, VersionedDao}
+import uk.ac.wellcome.dynamo.UpdateExpressionGenerator
 import uk.ac.wellcome.models._
 import uk.ac.wellcome.models.aws.S3Config
+import uk.ac.wellcome.storage.dynamo.VersionedDao
 import uk.ac.wellcome.storage.type_classes.HybridRecordEnricher
 import uk.ac.wellcome.s3.{S3ObjectLocation, S3ObjectStore}
 import uk.ac.wellcome.type_classes.{IdGetter, VersionGetter, VersionUpdater}
