+++ sbt_common/storage/src/test/scala/uk/ac/wellcome/storage/test/fixtures/LocalVersionedHybridStore.scala
@@ -4,10 +4,10 @@ import com.gu.scanamo._
 import com.gu.scanamo.syntax._
 import io.circe.Encoder
 import org.scalatest.{Assertion, Matchers}
-import uk.ac.wellcome.dynamo.VersionedDao
 import uk.ac.wellcome.models.aws.{DynamoConfig, S3Config}
 import uk.ac.wellcome.storage.vhs.{HybridRecord, VersionedHybridStore}
 import uk.ac.wellcome.models.Id
+import uk.ac.wellcome.storage.dynamo.VersionedDao
 import uk.ac.wellcome.s3.{KeyPrefixGenerator, S3ObjectStore}
 import uk.ac.wellcome.test.utils.JsonTestUtil
 import uk.ac.wellcome.utils.JsonUtil._
