+++ sbt_common/storage/src/test/scala/uk/ac/wellcome/storage/vhs/VersionedHybridStoreTest.scala
@@ -3,7 +3,6 @@ package uk.ac.wellcome.storage.vhs
 import com.gu.scanamo.{DynamoFormat, Scanamo}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.dynamo.VersionedDao
 import uk.ac.wellcome.models.aws.DynamoConfig
 import uk.ac.wellcome.models.{Id, Sourced, Versioned}
 import uk.ac.wellcome.s3.{KeyPrefixGenerator, S3ObjectStore}
