+++ sierra_adapter/sierra_items_to_dynamo/src/main/scala/uk/ac/wellcome/platform/sierra_items_to_dynamo/services/DynamoInserter.scala
@@ -1,9 +1,9 @@
 package uk.ac.wellcome.platform.sierra_items_to_dynamo.services
 
 import com.google.inject.Inject
-import uk.ac.wellcome.dynamo.VersionedDao
 import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
 import uk.ac.wellcome.platform.sierra_items_to_dynamo.merger.SierraItemRecordMerger
+import uk.ac.wellcome.storage.dynamo.VersionedDao
 import uk.ac.wellcome.utils.GlobalExecutionContext._
 
 import uk.ac.wellcome.dynamo._
