+++ sierra_adapter/sierra_items_to_dynamo/src/test/scala/uk/ac/wellcome/platform/sierra_items_to_dynamo/fixtures/DynamoInserterFixture.scala
@@ -1,10 +1,10 @@
 package uk.ac.wellcome.platform.sierra_items_to_dynamo.fixtures
 
 import com.gu.scanamo.DynamoFormat
-import uk.ac.wellcome.dynamo.VersionedDao
 import uk.ac.wellcome.models.aws.DynamoConfig
 import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
 import uk.ac.wellcome.platform.sierra_items_to_dynamo.services.DynamoInserter
+import uk.ac.wellcome.storage.dynamo.VersionedDao
 import uk.ac.wellcome.test.fixtures.{LocalDynamoDb, TestWith}
 import uk.ac.wellcome.test.fixtures.LocalDynamoDb.Table
